[modify] move MoreViewController's UITableViewDataSource to MoreDataSource.
